[Refactor/#21] 전체 도메인 id 자료형, Long으로 수정

+++ server/src/main/java/coumo/server/domain/Store.java
@@ -3,7 +3,6 @@ package coumo.server.domain;
 import coumo.server.domain.common.BaseEntity;
 import coumo.server.domain.enums.StoreType;
 import coumo.server.domain.mapping.CustomerStore;
-import coumo.server.web.dto.StoreRequestDTO;
 import jakarta.persistence.*;
 import lombok.*;
 import org.locationtech.jts.geom.Coordinate;
@@ -27,7 +26,7 @@ import java.util.List;
 public class Store extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "owner_id")
